app/src/main/java/com/example/composetask/viewmodels/HomeViewModel.kt
package com.example.composetask.viewmodels

import androidx.lifecycle.SavedStateHandle
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.example.composetask.models.Medicine
import com.example.composetask.repository.MedicineRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class HomeViewModel @Inject constructor(
    private val repository: MedicineRepository,
    val savedStateHandle: SavedStateHandle
) : ViewModel() {

    private val _medicines = MutableStateFlow<List<Medicine>>(emptyList())
    val medicines: StateFlow<List<Medicine>>
        get() = _medicines

    private val _isLoading = MutableStateFlow(true)
    val isLoading: StateFlow<Boolean>
        get() = _isLoading

    init {
        viewModelScope.launch {
            _isLoading.value = true
            repository.getMedicines().collect { medicineList ->
                _medicines.value = medicineList
                _isLoading.value = false
            }
        }
    }
}
